Added cloud bridge
